requirements.txt
ansible==7.1.0
ansible-core==2.14.1
ansible-vault==2.1.0
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.3
django-environ==0.9.0
django-filter==22.1
django-pytest==0.2.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
environ==1.0
exceptiongroup==1.0.1
factory-boy==3.2.1
Faker==15.3.1
future==0.18.2
generics==6.0.0
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mypy==0.990
mypy-extensions==0.4.3
oauthlib==3.2.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.2
PyJWT==2.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.2.0
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.4
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
resolvelib==0.8.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
